event: Only compare lowest byte of `__undesired`

+++ mcfgthread/event.c
@@ -24,7 +24,7 @@ _MCF_event_await_change_slow(_MCF_event* event, int undesired, const int64_t* ti
   try_wait_loop:
     _MCF_atomic_load_pptr_acq(&old, event);
     for(;;) {
-      if(old.__value != undesired)
+      if(old.__value != (uint8_t) undesired)
         return old.__value;
 
 #pragma GCC diagnostic push
